Refactor CMake, refactor folder
