+++ .idea/runConfigurations/OCD_FatShark_Power_V1.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="OCD FatShark_Power_V1" type="com.jetbrains.cidr.embedded.openocd.conf.type" factoryName="com.jetbrains.cidr.embedded.openocd.conf.factory" REDIRECT_INPUT="false" PASS_PARENT_ENVS_2="true" PROJECT_NAME="FatShark_Power_V1" TARGET_NAME="FatShark_Power_V1.elf" CONFIG_NAME="Debug" RUN_TARGET_PROJECT_NAME="FatShark_Power_V1" RUN_TARGET_NAME="FatShark_Power_V1.elf">
+  <configuration default="false" name="OCD FatShark_Power_V1" type="com.jetbrains.cidr.embedded.openocd.conf.type" factoryName="com.jetbrains.cidr.embedded.openocd.conf.factory" REDIRECT_INPUT="false" PASS_PARENT_ENVS_2="true" PROJECT_NAME="FatShark_Power_V1" TARGET_NAME="FatShark_Power_V1.elf" CONFIG_NAME="Flash 128k" RUN_TARGET_PROJECT_NAME="FatShark_Power_V1" RUN_TARGET_NAME="FatShark_Power_V1.elf">
     <openocd gdb-port="3333" telnet-port="4444" board-config="$PROJECT_DIR$/stmf1xx-jlink.cfg" reset-type="INIT" download-type="ALWAYS" />
     <method v="2">
       <option name="com.jetbrains.cidr.execution.CidrBuildBeforeRunTaskProvider$BuildBeforeRunTask" enabled="true" />
